Move functions from misc.py to new file print_products.py
